Remove requirejs
