+++ templates/common/list.tmpl
@@ -1,5 +1,4 @@
 {{template "header"}}
-	<script data-main="/static/javascript/common.js" src="static/javascript/require.js"></script>
 	<script src="static/javascript/list.js"></script>
 	<link rel="stylesheet" type="text/css" href="static/css/sharelink.css">
 	<link rel="stylesheet" type="text/css" href="static/css/list.css">
